pip-requirements.txt
PyBrain==0.3
argparse==1.2.1
numpy==1.8.0
pypng==0.0.16
scipy==0.13.2
wsgiref==0.1.2
